src/vscn-mapper/requirements.txt
dnspython==2.2.1
pyflakes==2.5.0
python-dotenv==0.20.0
psycopg==3.1.8
psycopg-binary==3.1.8
fuzzywuzzy[speedup]==0.18.0